#18813 Site-Key dns format validationmust be Config optional
